Split classes into their own file
