+++ decorators/trim.rb
@@ -1,0 +1,6 @@
+require_relative 'decorator'
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.length > 10 ? "#{@nameable.correct_name.strip[0..9]}..." : @nameable.correct_name
+  end
+end
